+++ dashboard/app/@dashboard/[model]/page.tsx
@@ -1,12 +1,8 @@
-import { Payment } from "@/components/table/columns";
-import { DataTable } from "@/components/table/data-table";
+import { DataTable } from "@/components/table-builder";
+import { fetchRecords, generateColumns } from "@/utils/model";
 
-export default async function DemoPage({
-  params,
-}: {
-  params: { model: string };
-}) {
-  const data = await getData(params.model);
+export default async function Model({ params }: { params: { model: string } }) {
+  const data = await fetchRecords(params.model);
   const columns = await generateColumns(data);
 
   return (
@@ -14,22 +10,4 @@ export default async function DemoPage({
       <DataTable columns={columns} data={data} />
     </div>
   );
-}
-
-async function getData(model: string): Promise<Payment[]> {
-  const data = await fetch(`http://localhost:8000/admin/models/${model}`, {
-    cache: "no-cache",
-  });
-  return data.json();
-}
-
-async function generateColumns(data: object[]) {
-  if (data && data.length > 0) {
-    return Object.keys(data[0]).map((key) => ({
-      accessorKey: key,
-      header: key.charAt(0).toUpperCase() + key.slice(1),
-    }));
-  } else {
-    return [];
-  }
 }
