+++ dashboard/app/@dashboard/layout.tsx
@@ -1,6 +1,6 @@
 import Sidebar from "./@sidebar/sidebar";
 export default function DashboardLayout({
-  children, // will be a page or nested layout
+  children,
 }: {
   children: React.ReactNode;
 }) {
