+++ dashboard/utils/sidebar.tsx
@@ -1,0 +1,8 @@
+async function fetchSideBar(): Promise<any[]> {
+  const data = await fetch(`http://localhost:8000/admin/models`, {
+    cache: "no-store",
+  });
+  return data.json();
+}
+
+export { fetchSideBar };
